+++ CHANGELOG.md
@@ -22,7 +22,7 @@
 ### Features
 
 * add HTML PR review rules ([8a79446](https://github.com/opsdevcode/opdevcode-website/commit/8a794463afae48a1ee67d06439e3b0d48887e472))
-* add initial OpsDevCo website ([ccb3e71](https://github.com/opsdevcode/opdevcode-website/commit/ccb3e7116d933bac93a6d3671b6fc1b73ffc4e77))
+* add initial OpsDevCode website ([ccb3e71](https://github.com/opsdevcode/opdevcode-website/commit/ccb3e7116d933bac93a6d3671b6fc1b73ffc4e77))
 * add logo options for feedback ([b8a1157](https://github.com/opsdevcode/opdevcode-website/commit/b8a1157b9f856ad6313b8355db2c8cc547b9e20f))
 * add logo options for feedback ([5582a19](https://github.com/opsdevcode/opdevcode-website/commit/5582a195de90913ea677e8c30d7aefee7db1f6c8))
 * add next-level site improvements ([4364e28](https://github.com/opsdevcode/opdevcode-website/commit/4364e28411593d556222566940bf8d6d9cc06df1))
@@ -66,6 +66,6 @@
 ### Features
 
 * add HTML PR review rules ([8a79446](https://github.com/opsdevcode/opdevcode-website/commit/8a794463afae48a1ee67d06439e3b0d48887e472))
-* add initial OpsDevCo website ([ccb3e71](https://github.com/opsdevcode/opdevcode-website/commit/ccb3e7116d933bac93a6d3671b6fc1b73ffc4e77))
+* add initial OpsDevCode website ([ccb3e71](https://github.com/opsdevcode/opdevcode-website/commit/ccb3e7116d933bac93a6d3671b6fc1b73ffc4e77))
 * add preview deploy workflow and CLI setup ([da8a134](https://github.com/opsdevcode/opdevcode-website/commit/da8a134321f537b30ac106b55406c50b4ec07869))
 * add preview deploy workflow and CLI setup ([0a8e9d3](https://github.com/opsdevcode/opdevcode-website/commit/0a8e9d34a11e6cdbf6e2997f41916fb609438e99))